feat(asset): Asset is no longer involved in upload
