+++ module/asset/src/types.ts
@@ -11,14 +11,3 @@ export interface Asset extends StreamMeta {
   source: Readable | string | Buffer;
   localFile?: string;
 }
-
-/**
- * An asset response
- */
-export interface AssetResponse extends StreamMeta {
-  stream(): Readable;
-  /**
-   * Response byte range, inclusive
-   */
-  range?: [start: number, end: number];
-}
