+++ module/asset/src/util.ts
@@ -8,10 +8,9 @@ import { getExtension, getType } from 'mime';
 
 import { path } from '@travetto/manifest';
 import { StreamMeta } from '@travetto/model';
-import { AppError } from '@travetto/base';
+import { StreamUtil } from '@travetto/base';
 
 import { Asset } from './types';
-import { LocalFile } from './file';
 
 /**
  * Utilities for processing assets
@@ -29,40 +28,11 @@ export class AssetUtil {
   }
 
   /**
-   * Read a chunk from a file, primarily used for mime detection
-   */
-  static async readChunk(input: string | Readable | Buffer, bytes: number): Promise<Buffer> {
-    if (Buffer.isBuffer(input)) {
-      return input;
-    } else if (typeof input === 'string') {
-      const fd = await fs.open(input, 'r');
-      try {
-        const buffer = Buffer.alloc(bytes);
-        await fd.read(buffer, 0, bytes, 0);
-        return buffer;
-      } finally {
-        try { fd.close(); } catch { }
-      }
-    } else {
-      const chunks: Buffer[] = [];
-      let size = 0;
-      for await (const chunk of input) {
-        const bChunk = Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk);
-        chunks.push(bChunk);
-        if ((size += bChunk.length) >= bytes) {
-          break;
-        }
-      }
-      return Buffer.concat(chunks);
-    }
-  }
-
-  /**
    * Detect file type from location on disk
    */
   static async detectFileType(input: string | Buffer | Readable): Promise<{ ext: string, mime: string } | undefined> {
     const { default: fileType } = await import('file-type');
-    const buffer = await this.readChunk(input, 4100);
+    const buffer = await StreamUtil.readChunk(input, 4100);
     const matched = await fileType.fromBuffer(buffer);
     if (typeof input === 'string' && matched?.mime === 'video/mp4' && input.endsWith('.m4a')) {
       return { ext: '.m4a', mime: 'audio/mpeg' };
@@ -138,19 +108,6 @@ export class AssetUtil {
   }
 
   /**
-   * File to blob
-   * @param file
-   * @param metadata
-   */
-  static async fileToBlob(file: string, metadata: Partial<StreamMeta> = {}): Promise<File> {
-    let type = metadata.contentType;
-    if (!type) {
-      type = (await this.detectFileType(file))?.mime;
-    }
-    return new LocalFile(file, type);
-  }
-
-  /**
    * Convert blob to asset structure
    */
   static async blobToAsset(blob: Blob, metadata: Partial<StreamMeta> = {}): Promise<Asset> {
@@ -188,49 +145,11 @@ export class AssetUtil {
   }
 
   /**
-   * Fetch bytes from a url
-   */
-  static async fetchBytes(url: string, byteLimit: number = -1): Promise<Buffer> {
-    const str = await fetch(url, {
-      headers: (byteLimit > 0) ? {
-        Range: `0-${byteLimit - 1}`
-      } : {}
-    });
-
-    if (!str.ok) {
-      throw new AppError('Invalid url for hashing', 'data');
-    }
-
-    let count = 0;
-    const buffer: Buffer[] = [];
-
-    for await (const chunk of Readable.fromWeb(str.body!)) {
-      if (Buffer.isBuffer(chunk)) {
-        buffer.push(chunk);
-        count += chunk.length;
-      } else if (typeof chunk === 'string') {
-        buffer.push(Buffer.from(chunk));
-        count += chunk.length;
-      }
-
-      if (count > byteLimit && byteLimit > 0) {
-        break;
-      }
-    }
-
-    try {
-      await str.body?.cancel();
-    } catch { }
-
-    return Buffer.concat(buffer, byteLimit <= 0 ? undefined : byteLimit);
-  }
-
-  /**
    * Compute hash from a url
    */
   static async hashUrl(url: string, byteLimit = -1): Promise<string> {
     const hasher = crypto.createHash('sha256').setEncoding('hex');
-    const finalData = await this.fetchBytes(url, byteLimit);
+    const finalData = await StreamUtil.fetchBytes(url, byteLimit);
     return hasher.update(finalData).end().read().toString();
   }
 }
