+++ module/base/test/fixtures/asset.yml
@@ -1,0 +1,3 @@
+asset:
+  yaml:
+    value: 10
