+++ module/rest-upload/package.json
@@ -22,13 +22,14 @@
   "main": "__index__.ts",
   "repository": {
     "url": "https://github.com/travetto/travetto.git",
-    "directory": "module/asset-rest"
+    "directory": "module/rest-upload"
   },
   "dependencies": {
     "@fastify/busboy": "^2.1.1",
     "@travetto/asset": "^4.1.3",
     "@travetto/config": "^4.1.2",
-    "@travetto/rest": "^4.1.4"
+    "@travetto/rest": "^4.1.4",
+    "file-type": "^16.5.4"
   },
   "peerDependencies": {
     "@travetto/test": "^4.1.1"
@@ -39,7 +40,7 @@
     }
   },
   "travetto": {
-    "displayName": "Asset Rest Support"
+    "displayName": "Rest Upload Support"
   },
   "private": false,
   "publishConfig": {
